Implement scrittore module
Scrittore takes care of the 'view' portion of the MVC architecture. This commit
implements the basic functionality of parsing note streams and calling the
requisite _annotation methods on the Groupings to ensure that the note streams
are interpolated with the current annotations (such as bar-checks, measure
comments, etc.). Also implements a first step toward Handlebars template
functionality and Serde.